:sparkles: divide main.ts
