+++ src/main.ts
@@ -2,9 +2,9 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ValidationPipe } from '@nestjs/common';
 import { ConfigService } from './config/config.service';
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 
 import * as CookieParser from 'cookie-parser';
+import { configure } from './config.main';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -15,16 +15,8 @@ async function bootstrap() {
     credentials: true,
   });
   app.use(CookieParser('secret'));
-  const configService = app.get(ConfigService);
-
-  const options = new DocumentBuilder()
-    .setTitle('Lokalne dobrodziejstwa')
-    .setDescription('API descriptions')
-    .setVersion('1.0')
-    .build();
-  const document = SwaggerModule.createDocument(app, options);
-  SwaggerModule.setup('api', app, document);
-
+  const configService = app.get<ConfigService>(ConfigService);
+  configure(app)
   await app.listen(configService.get('PORT'));
 }
